Replace the Branch option with the Variant
Initially the "Branch" option was designed as a way to easily switch
between stable, beta and other update channels.

However the current atomic update implementation doesn't have a concept
of branches. Instead, it currently uses the "variant" for this purpose.

With this commit we uniform this API to use the, currently standard,
"variant" value.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>