Bump meson version
Since meson version 0.56.0, "get_pkgconfig_variable()" has been
deprecated. With this commit we use the new "get_variable()" instead,
to avoid the deprecation warning.

Also Meson 0.58.0 includes the ability to append and prepend multiple
times to the same environment variable object. This will be useful
when setting the environment for the automated tests.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>