Store update JSON file in a well known location
Instead of storing the update JSON file in `/tmp` with a randomly
generated filename, we use a well known location under `/run`.

This gives us the advantage of being able to recover an update file even
if the atomupd-service is restarted.

`/run` is expected to be automatically cleaned on reboot, so there is no
need to manually delete the update file when it's not needed anymore.

If in the filesystem we have a malformed JSON, we expect the external
script `steamos-atomupd-client` to fail, and we will set the update
status to `AU_UPDATE_STATUS_FAILED`.
To overwrite the JSON file, it is enough to call the method
`CheckForUpdates`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>