tests: Store the mock RAUC service pid in a file
Instead of storing the mock RAUC service pid in an environment variable,
we use a text file. This will make it easier to pass around that
information from the parent to the child, even if we launch the service
when the child is already running.

We will start to use that in a following commit.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>