Restore the update status at start up
When an update has been successfully installed we store its buildid in a
text file called "/run/steamos-atomupd/reboot_for_update", indicating that
we are waiting for a reboot to complete the OS update.

If the daemon is restarted or crashes, we can use that file to recover
the update status.
Additionally we also stop any eventual steamos-atomupd-client and or
rauc processes that might be left running in background to start in a
clean and predictable state.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>