Do not propose updates that are already applied
If we apply an update and, without rebooting, we query again the server,
it will likely propose the same update a second time.

The client in theory could infer that the proposed update has already
been installed by checking that the `UpdateStatus` property is equal to
`3` (i.e. `SUCCESSFUL`) and the `UpdateVersion` is equal to the image
version that the server is proposing.

However it's better to keep most of the logic in one place and avoid
relying on the clients to do additional sanity checks.
For this reason, when this happens, we should just expose in the API
that no additional updates are available.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>