Add SwitchToVariant method instead of overloading CheckForUpdates
CheckForUpdates was used to check for available updates and also to
select the desired update variant.

This made the behavior different from what we currently have with the
CLI tools steamos-select-branch and steamos-update, where the variant
selection is a separate step from checking for updates.

Additionally, atomupd-daemon was storing the chosen variant nowhere.
Meaning that the clients had to manually remember their selection.

With this commit instead we move the variant selection from
CheckForUpdates to a dedicated SwitchToVariant method.
This new method will store the variant in the same text file as
steamos-select-branch does, to ensure cross-compatibility between the
two.
If you switch variant with steamos-select-branch, your new selection
will be automatically reflected in the "Variant" property too.

Also the extended/contracted versions of the Jupiter variants (e.g.
steamdeck->rel) will be correctly handled to ensure
steamos-select-branch/steamos-update backward compatibility.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>