Define the Atomupd1 interface
The interface was the last remaining piece for the D-Bus API.

As for the bus name and path, we don't currently expect the need to
change its value at build time. However defining it in meson will ensure
that we can use it in the daemon implementation and automated tests
without the need to add a dependency between each other.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>