Add a CLI atomupd-manager
Introduce a CLI wrapper to make it easier for developers to interact
with the daemon. The development of a CLI wrapper was in the initial
daemon API design as a nice to have feature.

This is an utility that is aimed only at developers and is expected to
be installed on Jupiter/Holo by default.
Clients (e.g. Steam or Discover) are always expected to rely upon the
canonical D-Bus API.

Once the clients start to use the D-Bus API, `atomupd-manager`
should be preferred by developers over the current legacy bash scripts
`steamos-update` and `steamos-select-branch`, because this CLI utilizes
the same API as the daemon does.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>