Add support for password protected test servers
When a new OS build is created, especially if it has potentially
breaking changes, it might be desirable to ask other developers/QA to
test it before publicly releasing it in the "main" branch.

This is due to the fact that "main" is still accessible by end users, so
images there are still expected to have a certain amount of stability.

By adding support for HTTP authentication, new risky changes can be
deployed to temporary test servers and that will allow developers/QA to
quickly test those new images before releasing them to the "main"
branch.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>