meson: Install atomupd-manager in the default bin directory
`atomupd-manager` is a CLI executable that eases the interaction with
the daemon.

This should be installed in the default `bin` directory and not in
`/usr/lib/`, because developers are supposed to find it in their PATH.

It is safe to move this executable because at the moment there are no
other scripts that depend/use it.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>