Validate buildid of the current system
When starting up the atomupd daemon, we validate the current system
buildid. The check is conceptually the same of to what it's done in the
BuildId class of `steamos-atomupd.git`.

If the buildid is unexpected, this is a clear indication that something
is wrong with the manifest or the whole system.
In that case we exit early with an error.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>