Add polkit policy to check for updates
With polkit we can limit the users that can interact with the API
without having to authenticate with an admin user first.

Given the security implications of searching and applying system
updates, we decided to require admin authentication for all users that
are not in "active sessions".

We start by limiting "CheckForUpdates" and with the following commits
we'll cover the other methods too.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>