Add polkit rule for SwitchToVariant method
Switching to a different variant is inherently a risky operation when
compared to, for example, checking for available updates.

To mitigate this risk we set a policy that requires all users to
authenticate with admin privileges before they can utilize the
SwitchToVariant method.

However, in order to still allow the Steam client on the Steam Deck to
fully interact with the API, we add a unique exception rule for the
"deck" user.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>