Add polkit rule for StartUpdate
Based on the requested buildid, we make a distinction between upgrades
and downgrades.

An upgrade is a relatively safe operation, so we allow every "active"
user to perform it.

On the other hand, a downgrade is riskier and could even be used for a
downgrade attack.
For this reason, similarly to "SwitchToVariant", we require all users to
authenticate with admin privileges, and we keep the "deck" user as the
only exception to this rule.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>