Do not overload the word "version", when we are referring to build IDs
When this API was initially designed, all Jupiter images were
"snapshots".

Recently we started to create properly versioned images.
This change has caused some confusion regarding how to interpret the
meaning of "version".

In the image manifests we have a "version" field, which represents a
concept similar to a branch. For example we might have an image with a
version "3.4.9" and a build ID "20230901.1". Another image might also
have the same version "3.4.9" but a different build ID, such as
"20230902.1".

Instead, in this API, we originally used "version" to uniquely identify
a particular image update.

With this commit we align the API's naming scheme with what we use in
the image JSON manifest file.
This means that, for example, `CurrentVersion` now holds the image
"version", as the manifest does. Also the build ID has been moved to a
separate property called `CurrentBuildID`.

Technically this is an API breaking change that would require a D-Bus
interface version bump.
However, at the moment only KDE Discover is actively utilizing this API,
and these changes don't have a disrupting impact on its implementation.
The only property affected by this commit, that Discover accesses, is
`CurrentVersion`. Discover uses that value only to show users what's
the upgrade path, and for checking if the daemon is up and running.
This means that if we release a new atomupd-daemon version without a fix
for Discover, users will potentially see the system update as e.g.
"3.5.0 -> 20230831.1000".
Given that we control the packages version in Jupiter, we can skip an
interface bump and just increase the "Version" property instead.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>