atomupd-manager: Use custom log handler
g_log_default_handler() by default drops debug and info logs.
This means that even if we use
`g_log_set_handler(..., G_LOG_LEVEL_DEBUG, g_log_default_handler, ...)`,
the debug messages will not be printed in output.

With this commit we create a custom log handler that allows us to
include debug and info messages when launching the CLI with the option
`--verbose`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>