atomupd-manager: Enable daemon debug mode when using --verbose
The majority of informative logs are provided by the atomupd-daemon
D-Bus API.
For this reason, when executing atomupd-manger with the `--verbose`
option, we also enable the debug mode for atomupd-daemon and parse its
logs from the journal.

This gives developers a convenient option to gather a more complete view
of what is happening.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>