Change --config option to --config-directory
Instead of letting users re-start the daemon with a specific custom
configuration file, we allow them to choose a different directory where
the daemon will look for the client configuration file.

This is in preparation of adding an API property to customize how to
treat the eventual `client-dev.conf` file.

It is safe to change the `--config` option because so far it was only
used in the automated tests.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>