atomupd-manager: Increase the dbus message timeout
When using the method CheckForUpdates, the API downloads the meta JSON
file and then replies with its parsed content.
That operation usually only takes a few ms to complete, but on very slow
Internet connections it could take seconds.
For this reason we increase the timeout to 20 seconds, which should be
enough to differentiate between a slow connection to a stale one.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>