utils: Support up to two additional sub dirs in Desync URL config
The server doesn't have a strict rule regarding how to organize the
images. At the moment the castr path has always been three directories
down compared to the root, but this doesn't mean that in the future we
might want to add an additional directory to better organize them.

With this commit we start to create the Desync config file with support
for up to five sub directories, which should give us enough room for
manoeuvre.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>