Expect the image manifest to have a default branch value
Since SteamOS 3.6 all image manifests are expected to have the
`default_update_branch` value.

There is no need to keep a fallback to support legacy images because it
is unlikely that we'll ever need to update atomupd-daemon in the old
SteamOS 3.5 images, where the variant concept changed into variant plus
branch.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>