Use the new default_update_branch property from the manifest
In the past the `branch` field inside an image manifest was used as the
default fallback choice, until the user manually selected which channel
to follow.

However, in order to support more complex use cases, the `branch` field
was superseded by the new `default_update_branch`.

More detailed information are available in the steamos-atomupd
documentation.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>