Use `stable` as the last resort branch
When we don't have a user preferences file and the image manifest is
lacking the expected default branch value, we can do a final attempt at
using `stable` as an hardcoded fallback branch.

This is a last resort fallback, before marking the configuration as
borked and requiring some manual intervention to fix it.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>