Use is_parallel option to prevent running tests concurrently
The majority of the automated tests require holding a D-Bus name, which
breaks if multiple tests are executed concurrently.

Instead of relying on running meson with `--num-processes 1`, we
directly declare them with `is_parallel: false`.

This has the same end result, but has the advantage that we can simply
run `meson test` without manually specifying additional launch options
every time.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>