Try to continue even if the config/preferences files are malformed
If the legacy configuration file is present, but malformed, we can
ignore it and try to fallback to the preferences file.
If also the preference file is malformed, we can attempt to parse the
image manifest file as a last resort.

This change makes the atomupd daemon more resilient against unexpected
configuration files. It is better to start the daemon with a potentially
different branch/variant, compared to what the users chose the last
time, instead of hard failing and requiring manual intervention to
unblock it.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>