Revert to the default variant and branch when receiving an HTTP 4xx error
If the query server returns an HTTP error in the 4xx range, e.g. 404 or
401, then this is symptom that our request was incorrect.

When this happen, we try to revert the selected variant and branch to
the default values. This is a last resort attempt at automatically
fixing the atomic update process without requiring any manual
intervention by the end users.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>