Split _au_parse_config into separated functions
Parse the preferences and manifest separately from the atomupd
configuration.

In a followup commit this will allow us to preload the necessary
information to download the new remote info file.

This is mostly a refactoring. The only function difference is that now
`ReloadConfiguration` will only reload the configuration, and not also
the preferences and manifest.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>