Add support for a server side "remote-info.conf" file
When starting up the atomupd-daemon, we try to download the new
"remote-info.conf" file from the server.
That file can be used to provide a more up-to-date list of known
branches and variants.

For example like the recent addition of the "preview" and "pc" branches.

Clients will be able to download that file, if available, and replace
their potentially outdated branches/variants knowledge.

The download happens only at start time because we expect that the
"remote-info.conf" file should be updated very rarely on the server
side. If the download fails, and we don't have already a local copy in
`/etc/steamos-atomupd/remote-info.conf`, we simply fallback to the
branches/variants info from the `client.conf` as usual.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>