au-atomupd1-impl: Ensure Desync config file to be present
The Desync config file in `/etc/desync/config.conf` should be provided
by the steamos-customization package, a dependency of atomupd-daemon.

If for whatever reson that file, or parent directory, is missing, we can
take care of creating an empty configuration as a last attempt to keep
the OS updates working.

Addresses https://gitlab.steamos.cloud/holo-team/tasks/-/issues/1714

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>