atomupd-manager: Increase the message timeout to 30 seconds
If the server is not providing the pre-estimated update size,
`steamos-atomupd` is gonna download a very small raucb file to do the
estimation locally.
That download operation has a timeout of 20 seconds, after which the
checking for updates process continues without the estimation part.

We increase the timeout in atomupd-manager to cover that scenario and
avoid returning early.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>